fix: address all Skyfall review feedback on PR #109
- Move inline dashboard icon to Icons.tsx (AlertTriangle)
- Fix banned.tsx: slack.hackclub.com link + remove slackId fallbacks
- Remove global stats from admin dashboard, fix name fallbacks, componentize admin views
- Restore blocked-prompts.ts entries (Codex CLI, Trae)
- Restore price.ts + add cost column back to activity
- Fix migration journal timestamps (realistic ms values)
- Consolidate banned fields: remove isBanned, use reviewStatus only
- Add local moderation classifier (regex classifier + async background moderation for all output
- Make moderation API keys optional, add guard to moderations proxy
