+++ drizzle/0020_drop_is_banned.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "users" DROP COLUMN IF EXISTS "is_banned";
