+++ drizzle/meta/_journal.json
@@ -117,29 +117,36 @@
     {
       "idx": 16,
       "version": "7",
-      "when": 1779302500000,
+      "when": 1779302501983,
       "tag": "0016_add_content_column",
       "breakpoints": true
     },
     {
       "idx": 17,
       "version": "7",
-      "when": 1779302600000,
+      "when": 1779302604157,
       "tag": "0017_add_violation_event_id",
       "breakpoints": true
     },
     {
       "idx": 18,
       "version": "7",
-      "when": 1779302700000,
+      "when": 1779302708291,
       "tag": "0018_add_violation_dismissed",
       "breakpoints": true
     },
     {
       "idx": 19,
       "version": "7",
-      "when": 1779400000000,
+      "when": 1779400003142,
       "tag": "0019_add_is_admin_to_users",
+      "breakpoints": true
+    },
+    {
+      "idx": 20,
+      "version": "7",
+      "when": 1779400018273,
+      "tag": "0020_drop_is_banned",
       "breakpoints": true
     }
   ]
