+++ src/db/schema.ts
@@ -31,7 +31,6 @@ export const users = pgTable(
     }),
     isIdvVerified: boolean("is_idv_verified").notNull().default(false),
     skipIdv: boolean("skip_idv").notNull().default(false),
-    isBanned: boolean("is_banned").notNull().default(false),
     isAdmin: boolean("is_admin").notNull().default(false),
     agentBannerDismissedAt: timestamp("agent_banner_dismissed_at"),
     reviewStatus: text("review_status").notNull().default("normal"), // "normal" | "flagged" | "strict" | "banned"
