+++ src/env.ts
@@ -9,8 +9,8 @@ const envSchema = type({
   SLACK_GEOBLOCK_WEBHOOK_URL: "string.url",
   OPENAI_API_URL: "string",
   OPENAI_API_KEY: "string",
-  OPENAI_MODERATION_API_KEY: "string",
-  OPENAI_MODERATION_API_URL: "string",
+  "OPENAI_MODERATION_API_KEY?": "string",
+  "OPENAI_MODERATION_API_URL?": "string",
   ALLOWED_LANGUAGE_MODELS: "string",
   ALLOWED_IMAGE_MODELS: "string",
   ALLOWED_EMBEDDING_MODELS: "string",
