+++ src/views/dashboard.tsx
@@ -4,6 +4,7 @@ import type { Stats, User } from "../types";
 import { AgentBanner } from "./components/AgentBanner";
 import { Header } from "./components/Header";
 import { IdvBanner } from "./components/IdvBanner";
+import { AlertTriangle } from "./components/Icons";
 import { StatCard } from "./components/StatCard";
 import { Layout } from "./layout";
 
@@ -136,22 +137,7 @@ export const Dashboard = ({
               href="/dashboard/violations"
               title="Violations"
               description="View your content violations and review status."
-              icon={
-                <svg
-                  aria-hidden="true"
-                  class="w-8 h-8"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                  stroke-width="2"
-                >
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-                  />
-                </svg>
-              }
+              icon={<AlertTriangle class="w-8 h-8" aria-hidden />}
             />
             {replicateEnabled && (
               <QuickLinkCard
